Refactor: Move Meraki API to its own directory
This commit moves the Meraki API code out of the `api` directory and into its own `meraki_api` directory. This is the first step in a larger refactoring to improve the maintainability of the codebase. I was unable to remove the old `meraki_api` directory due to an error, but this commit contains the new directory and the moved files.
